+++ source/parameters/uniform/uniform.h
@@ -10,10 +10,10 @@ class Uniform : public Korali::Parameter::Base
 {
  private:
    double _aux;
-   
+
  public:
 
-	 Uniform(nlohmann::json& js, int seed);
+	 Uniform(nlohmann::json& js, size_t seed);
 	 Uniform(double _lowerBound, double _upperBound, size_t seed = 0);
 	 ~Uniform();
